fix(cms): Add developemt environment condition + Move logic users seed logic to seperate file
