+++ package.json
@@ -11,7 +11,11 @@
   "devDependencies": {
     "@commitlint/cli": "^19.1.0",
     "@commitlint/config-conventional": "^19.1.0",
+    "@types/node": "^20.12.11",
     "husky": "^9.0.11",
     "lint-staged": "^15.2.2"
+  },
+  "dependencies": {
+    "@strapi/strapi": "^4.24.2"
   }
 }
